+++ app/components/layout/feedback.tsx
@@ -2,8 +2,9 @@
 
 import { useBreakpoint } from "@/app/hooks/use-breakpoint"
 import { FeedbackForm } from "@/components/common/feedback-form"
-import { Dialog, DialogContent, DialogTrigger } from "@/components/ui/dialog"
-import { Drawer, DrawerContent, DrawerTrigger } from "@/components/ui/drawer"
+import { Dialog, DialogContent, DialogTitle, DialogTrigger } from "@/components/ui/dialog"
+import { Drawer, DrawerContent, DrawerTitle, DrawerTrigger } from "@/components/ui/drawer"
+import { VisuallyHidden } from "@/components/ui/visually-hidden"
 import { useState } from "react"
 
 type FeedbackProps = {
@@ -25,6 +26,9 @@ export function Feedback({ trigger, authUserId }: FeedbackProps) {
         <Drawer open={isOpen} onOpenChange={setIsOpen}>
           <DrawerTrigger asChild>{trigger}</DrawerTrigger>
           <DrawerContent className="bg-background border-border">
+            <DrawerTitle>
+              <VisuallyHidden>Feedback Form</VisuallyHidden>
+            </DrawerTitle>
             <FeedbackForm authUserId={authUserId} onClose={handleClose} />
           </DrawerContent>
         </Drawer>
@@ -37,6 +41,9 @@ export function Feedback({ trigger, authUserId }: FeedbackProps) {
       <Dialog open={isOpen} onOpenChange={setIsOpen}>
         <DialogTrigger asChild>{trigger}</DialogTrigger>
         <DialogContent className="[&>button:last-child]:bg-background overflow-hidden p-0 shadow-xs sm:max-w-md [&>button:last-child]:top-3.5 [&>button:last-child]:right-3 [&>button:last-child]:rounded-full [&>button:last-child]:p-1">
+          <DialogTitle>
+            <VisuallyHidden>Feedback Form</VisuallyHidden>
+          </DialogTitle>
           <FeedbackForm authUserId={authUserId} onClose={handleClose} />
         </DialogContent>
       </Dialog>