+++ app/page.tsx
@@ -1,7 +1,7 @@
 import Chat from "./components/chat/chat"
 import LayoutApp from "./components/layout/layout-app"
 
-export default async function Home() {
+export default function Home() {
   return (
     <LayoutApp>
       <Chat />