+++ lib/chat-store/types.ts
@@ -1,5 +1,17 @@
-import type { Tables } from "@/app/types/database.types"
+import { Tables } from "@/app/types/database.types"
 
-export type Chat = Tables<"chats">
-export type Message = Tables<"messages">
-export type ChatHistory = Pick<Chat, "id" | "title" | "created_at">
+export type DBChat = Tables<"chats">
+export type DBMessage = Tables<"messages">
+
+export interface ChatHistory {
+  id: string
+  title: string
+  created_at: string
+  _deleted?: boolean
+}
+
+export interface Chat extends ChatHistory {
+  user_id: string
+  model: string
+  system_prompt: string
+}